#3 skeleton
I spent the entire night trying to figure out how to do the fancy
iMessage drag down to dismiss to no avail, I found an anchored chat box
that I like and I after messing with it for a long time I decided to
say F U C K I T and just made it so when you tap the chat window itself
the keyboard is dismissed.

The problem with all the solutions I came up with for fancy drag swipe
was that the anchored uiview on top of the keyboard would never animate
WITH the keyboard, always after it, like disconnect, floating in space,
then race to the bottom. fuck ittttt so annoying. Anyway. Got the
skeleton of chats up :)